% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_ccast.R
\name{bar_plot_ccast}
\alias{bar_plot_ccast}
\title{Plotting Barplots for Homogeneous subpopulations}
\usage{
bar_plot_ccast(M, M2, file)
}
\arguments{
\item{M}{mxn matrix of means for m subpopulations and n markers}

\item{M2}{mxn matrix of standard deviations for m subpopulations and n
markers}

\item{file}{name of pdf output file to be saved in the current directory}
}
\value{
A pdf file labeled "Barplot for homogenous cells.pdf" with
barplots for all homogeneous subgroups.
}
\description{
This function takes as input 2 matrices corresponding to means and
standard deviations respectively of dimension mxn each across all markers
of size n and produces a pdf file of barplots for all subpopulation of
size m corresponding to the leaf of the ccast tree.
}
\examples{
bar_plot_ccast(M=M, M2=M2, "Barplot for homogenous cells.pdf")

}
\seealso{
\code{\link{ccast_filter5}} which this function wraps
}
