% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccast_biaxialplot.R
\name{ccast_biaxialplot}
\alias{ccast_biaxialplot}
\title{Plotting Biaxial plots for all inner nodes of ccast decision tree output}
\usage{
ccast_biaxialplot(optccastreeoutput, ylabel)
}
\arguments{
\item{optccastreeoutput}{An optimal ccast tree object produced by the function ccast_optimize}

\item{ylabel}{Any biomarker label provided by the user preferably one that is not used to build the tree}
}
\value{
A pdf file labeled "Biaxial tree node plots.pdf" with biaxial plots for all inner nodes from the optimal decision tree.
}
\description{
This function takes as input an optimal ccast decision tree output from the ccast_optimize function and any biomarker label not used in building the tree provided by the user and produces a pdf file with all pairwise scatter plots of all the inner nodes in the decision tree with density contours and split point cut off estimates used in building the tree.
}
\examples{
ylabel<-"M1"
data(optccastree)
ccast_biaxialplot(optccastreeoutput=optccastree,ylabel=ylabel)

}
\seealso{
\code{\link{ccast_filter5}} which this function wraps
}
