% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccast_heatmaplot.R
\name{ccast_heatmaplot}
\alias{ccast_heatmaplot}
\title{Plotting Heatmaps for Homogeneous subpopulations}
\usage{
ccast_heatmaplot(optccastreeoutput, file = "Heatmaps Homogeneous cells.pdf")
}
\arguments{
\item{optccastreeoutput}{An optimal ccast tree(decision tree) object produced by the function ccast_optimize}
}
\value{
an R objects corresponding to (1) a list of row ids(cell ids) (2) Tree node ids corresponding to all Subpopulations  and a pdf file labeled "Heatmaps Homogeneous cells.pdf" of heatmaps for all leaf nodes from the ccast tree. In addition to a list of 3 items:
\item{ccastmeans }{mxn matrix of means for m subpopulations and n markers }
\item{ccastsds }{mxn matrix of standard deviations for m subpopulations and n markers}
\item{idcells}{list of cell ids for all lineage subpopulations}
}
\description{
This function takes as input an optimal ccast decision tree output from the ccast_optimize function and produces 2 matrices corresponding to means and standard deviations respectively of dimension mxn each across all markers of size n and produces a pdf file of heatmaps for all subpopulation of size m.
}
\examples{
ccast_heatmaplot(optccastreeoutput=optccastree)

}
\seealso{
\code{\link{ccast_filter5}} which this function wraps
}
