% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccast_L.R
\name{ccast_L}
\alias{ccast_L}
\title{Determining the optimal prunning level L of the initial ccast tree}
\usage{
ccast_L(DD, mm)
}
\arguments{
\item{DD}{a dataframe of biomarker expression derived  from fcs files containing an extra column for cluster assignments labeled as "groups".}

\item{mm}{maximum number of predicted clusters}
}
\value{
A pdf file labeled "initial_tree.pdf" corresponding to the initial ccast decision tree. In addition a list of 3 items:
\item{treenodes }{list of data of all leaf nodes of the ccast tree corresponding to homogeneous subgroups}
\item{level }{maximum level L for prunning decision tree}
\item{initialtree}{initial ccast tree output with less homogeneous bins in the leaf nodes}
}
\description{
This function takes a dataframe of biomarkers and cluster assignments and produces a ccast decision tree with L levels
}
\examples{
ccastree<-ccast_L(dD1,mm)

}
\seealso{
\code{\link{ccast_filter5}} which this function wraps
}
