% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccast_optimize.R
\name{ccast_optimize}
\alias{ccast_optimize}
\title{Optimize the initial decision tree with height L for homogenous clusters.}
\usage{
ccast_optimize(t1, dD1, s)
}
\arguments{
\item{t1}{A tree object from the function ccast_L corresponding to homogeneous subgroups in the leaf nodes of the initial decision tree.}

\item{dD1}{a dataframe of biomarker expression derived  from fcs files containing an extra column for cluster assignments labeled as "groups".}

\item{s}{maximum level L for prunning decision tree}
}
\value{
A pdf file labeled "finaltree.pdf" corresponding to the final ccast decision tree. In addition a list of 3 items:
\item{tree }{final ccast tree output with distinct homogeneous bins in the leaf nodes.}
\item{fprediction }{list representing the distributions of events in each leaf node in the optimized ccast decision tree.}
\item{puredata}{list of all updated filtered data corresponding to each iteration in the optimization step. }
}
\description{
Optimize the initial decision tree with height L for homogenous clusters.
}
\section{description}{
 This function optimizes the decision tree by maximizing the size of the homogenous clusters.
}

\examples{
optccastree<-ccast_optimize(t1,dD1,s)

}
\seealso{
\code{\link{ccast_filter5}} which this function wraps
}
