% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccast_silhouette.R
\name{ccast_silhouette}
\alias{ccast_silhouette}
\title{Optimizing the density of cell subpopulations.}
\usage{
ccast_silhouette(d, fileout, nCluster.init = k, assign, iter.max = 10,
  diagnosis = TRUE)
}
\arguments{
\item{d}{distant measurement. Object  dist()}

\item{fileout}{pdf output file. Default labeled "silhouette.pdf"}

\item{nCluster.init}{number of initial clusters}

\item{assign}{cluster assignments from any clustering algorithm}

\item{iter.max}{max number of iteration}

\item{diagnosis}{logical for ploting silouette diagnostic plots. Default is TRUE}
}
\value{
A pdf file labeled "silhouette.pdf" corresponding to the silhouette diagnostic plots. In addition:
\item{orgFit }{Output from initial silhouette run}
\item{optFit }{Output from final silhouette run}
\item{optFitCluster }{Dataframe of final cluster assignments, final optimal number of clusters, optimal number of iteration and optimal cost function statistics}
}
\description{
This function uses Silhouette coefficients to optimize both the number of cell clusters as well as cell cluster distributions.
}
\examples{
assign <- cutree(hc, k=4)
out1=ccast_silhouette(d,fileout="silhouette.pdf", nCluster.init = 4, assign, iter.max = 10, diagnosis = TRUE)
}
\seealso{
\code{\link{ccast_filter5}} which this function wraps
}
