% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparepopintrasignalalivefidelity.R
\name{comparepopintrasignalalivefidelity}
\alias{comparepopintrasignalalivefidelity}
\title{A function to compare both lineage and intracellular signaling changes across pre defined survival cell populations.}
\usage{
comparepopintrasignalalivefidelity(rowids, dataP, asinhp, drugs, sigids, colid,
  antibody, patient)
}
\arguments{
\item{rowids}{List of row ids(cell ids) for each sub population of cells stored as R object.}

\item{dataP}{Marker expression data matrix of pooled cells from original FCS file stored as R object.}

\item{asinhp}{asinh parameter p for scaling the data if necessary. Default is 1.}

\item{drugs}{Parameter vector corresponding to the names of all single drugs.}

\item{sigids}{Vector of marker ids corresponding to the functional proteins used for building the Nested effects models.}

\item{colid}{column ids of lineage markers to use of equal length as colid. Must be provided by user.}

\item{antibody}{Vector of marker names of length equal to length of the columns in each FCS file for each channel in the FCS file.}

\item{patient}{Patient or subject name. Name must be included exactly as in the file names.}
}
\value{
List of dataframes with mean and fold change expression values for each drug relative to the baseline and heatmaps of fold changes with relative percentage of cells compared to basal for all drugs. The output of this functionis a list 5 items:
\item{singlecelldata}{List of dataframes for each subpopulation with drug response expression values for each cell across intracellular proteins }
\item{Popmeanexprsdata}{3D array for each subpopulation(3rd dimension) with mean expression for proteins(rows) and drugs (columns).}
\item{Diffpopmeanexprsdata}{3D array for each subpopulation(3rd dimension) with mean FC expression for lineage proteins(rows) and drugs (columns) relative to the basal treatment.}
\item{Difflineagepopmeanexprsdata}{3D array for each subpopulation(3rd dimension) with mean Fold change expression for lineage proteins(rows) and drugs (columns) relative to the basal treatment.}
\item{unstablepop}{Numeric vector corresponding to the lineage subpopulations likely affected by drugs.}
}
\description{
This function takes as input treatment response single-cell expression data labeled by cell types and returns  a list of dataframes and heatmaps of fold changes with relative percentage of cells compared to basal for all drugs.
}
\examples{
res2=comparepopintrasignalalivefidelity(rowids2,dataP,asinhp=asinhp,drugs=drugs,sigids=sigids,colid=colid,antibody=antibody,patient=patient)

}
\seealso{
\code{\link{drugmnemmain}} which this function wraps
}
