% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugnemmain.R
\name{drugnemmain}
\alias{drugnemmain}
\title{The main function for analyzing heterogeneous drug response data across a population of single cells to optimize combination therapy based on desired intracellular effects.}
\usage{
drugnemmain(patient, file, outputfile, manualgates = FALSE, celltype = NULL,
  drugs, drugids, antibody, pProteins, transformlogic = FALSE, asinhp = 1,
  colid = NULL, coln = NULL, rown = NULL, k = NULL, param = NULL,
  respondermarker = NULL, infer = "search", fcsdes = NULL,
  ylabel = ylabel, subanalysis = FALSE, subsamplesize = 20000,
  type = "CONTmLL", filter = TRUE, weighttype = "FC",
  targetprior = TRUE, Fidelity = FALSE, optimizenoeffect = FALSE)
}
\arguments{
\item{patient}{Patient or subject name for data. Name must be included exactly as in the FCS file names.}

\item{file}{A file path or directory path name with all FCS files of interest or a cxm flow cytometry expression matrix of c number of cells and m number of markers. The last 2 columns for the data matrix must be labeled  "treatment" and "groups" respectively corresponding to the index combination of drug label and cell cluster for each cell.}

\item{outputfile}{Name of output directory. drugnemain will set this as the working directory for storing all output files.}

\item{manualgates}{A logical indicating whether input folder contains manual gated data or not. Default is FALSE.}

\item{celltype}{Parameter vector corresponding to the manually gated cell types. Celltype names and ordering must be exact as in the manuall gated file names. Default is NULL if there are no manually gated data.}

\item{drugs}{Parameter vector corresponding to the names of all single drugs. Paremeter must maintain the order and names exactly as in the input folder with the FCS files with the Basal or control file name first.}

\item{drugids}{Integer vector of subset index of drug ids to be used for DRUGNEM analysis derived from the previous drugs parameter to optimize a subset of drugs.}

\item{antibody}{Vector of marker names of length equal to length of the columns in each FCS file for each channel in the FCS file.}

\item{pProteins}{Vector of markers corresponding to the functional proteins used for building the Nested effects models.}

\item{transformlogic}{An character string that defines the type of transformation. See function read.FCS in R packgae flowCore for more information. Valid values are linearize (default) or scale.The linearize transformation applies the appropriate power transform to the data while the scale transformation scales all columns to $[0,10^decades]$. defaulting to decades=0 as in the FCS4 specification. A logical can also be used: TRUE is equal to linearize and FALSE(which is the default) corresponds to no transformation.}

\item{asinhp}{asinh cofactor parameter p for scaling the data if necessary or NULL. Default is 1.}

\item{colid}{index of lineage markers of interest to be used by ccast algorithm to define cell clusters. Must be provided by user.}

\item{coln}{column names of lineage markers to use of equal length as colid. Must be provided by user.}

\item{rown}{row names of data to use of equal length as number of cells. If not provided the default FCS file names will be used.}

\item{k}{minimum number of expected clusters at baseline. Must be specified by user.}

\item{param}{string of parameter names from an FCS file. Default is NULL.}

\item{respondermarker}{Set of apoptotic or desired endpoint markers speficied by the user. Names must correspond to names in the input FCS files.}

\item{infer}{Network search approach to score the best drug network. Default is "search" for exhaustive search. For large networks(>5 nodes), we recommend "nem.greedy"}

\item{fcsdes}{Description of parameter names from an FCS file. Default is NULL.}

\item{ylabel}{y axis marker variable used to generate biaxial plots for all inner node marker variables from the optimized CCAST decision tree. Must be one of the names from the lineage marker list.}

\item{subanalysis}{A logical to downsample original FCS file(s). Default is FALSE. Recommended if total number of cells across all FCS files > 100000 to save time.}

\item{subsamplesize}{Number of cells to subsample using density downsampling approach by Qiu et al. 2011}

\item{type}{Parameter defining the type of data and likelihood for scoring the Nested effect models: Default "CONTmLL" called internally by the nem function in nem r-package correspond to the effect probabilities while "CONTmLLBayes" correspond to logodds that each protein is differentially expressed.}

\item{filter}{A logical to filter out death or dying cells or not. Default is TRUE.}

\item{weighttype}{parameters defining the type of desired phenotype information to use for optimizing treatment effects. Values include: (i) 'T-stat': Regulation using T-stat of intracellular signaling effects  (ii) 'FC'(default): Regulation based on Foldchange of intracellular signaling effects, and (iii) 'deathmarker': effects associated with regulation of a death effector protein or any other desired marker.}

\item{targetprior}{A logical to filter out outlier or noisy or alien drug targets. Default is TRUE.}

\item{Fidelity}{A logical to test for fidelity of pre vs post treatment characterisation of surface or lineage subpopulations using Simes statistics and identify potential unstable baseline subpopulations. Default is FALSE.}

\item{optimizenoeffect}{A logical to optmize drug combinations for No effect DRUG-NEM model. Useful for normal data. Default is FALSE.}
}
\value{
The output of this function is a list of the following 33 items.  Item 2 corresponds to the DRUGNEM ranking using "Any effect or no prior", item 10 corresponds to the DRUGNEM ranking using prior "Up regulation effects", item 12 correspond to the DRUGNEM ranking using prior "Down regulation effects", item 14 corresponds to the DRUGNEM ranking using prior "No effects" and item 33 corresponds the best ranking after comparing priors "Up effects" and "Down effects". It also generates several pdf(Tiff) files corresponding to various diagnostic plots including plots of initial and final optimized CCAST decision  trees for all cell  lineage sub groups, Heatmaps of fold changes for all functional or intracellular  markers across all sub populations and DRUGNEM network and ranking list of all drug combinations with corresponding score.
\item{DRUGNEMeffectmodel}{Graphical model showing the nested relationships between drugs and their target proteins using no prior weighting of desired effect data}
\item{Rankingeffectmodel}{Table showing the rankings and scores for all drug combinations using no prior desired drug effects.}
\item{DRUGNEMupmodel}{Graphical model showing the nested relationships between drugs and their target proteins using upregulation desired effect data.}
\item{Rankingupmodel}{Table showing the rankings and scores for all drug combination using upregulation desired effect data.}
\item{DRUGNEMdownmodel}{Graphical model showing the nested relationships between drugs and their target proteins using down regulation effect data.}
\item{Rankingdownmodel}{Table showing the rankings and scores for all drug combination using down regulation desired effect data.}
\item{DRUGNEMnoeffectmodel}{Graphical model showing the nested relationships between drugs and their target proteins using no effect desired effect data.}
\item{Rankingnoeffectmodel}{Table showing the rankings and scores for all drug combination using no effect desired effect data.}
\item{DRUGNEMupmarkermodel}{Graphical model showing the nested relationships between drugs and their target proteins using only upregulation desired effect data.}
\item{Rankingupmarkermodel}{Table showing the rankings and scores for all drug combination using only upregulation desired effect data.}
\item{DRUGNEMdownmarkermodel}{Graphical model showing the nested relationships between drugs and their target proteins using only down regulation desired effect data.}
\item{Rankingdownmarkermodel}{Table showing the rankings and scores for all drug combination using only down regulation desired effect data.}
\item{DRUGNEMnoeffectmarkermodel}{Graphical model showing the nested relationships between drugs and their target proteins using only no effect desired effect data.}
\item{Rankingnoeffectmarkermodel}{Table showing the rankings and scores for all drug combination using only no effect desired effect data.}
\item{DRUGNEMbesteffectmodel}{Graphical model showing the nested relationships between drugs and their target proteins using best desired effect data.}
\item{Rankingbesteffectmodel}{Table showing the rankings and scores for all drug combination using best desired effect data.}
\item{Up_object}{DRUGNEM R object output using upregulation desired priors.}
\item{Down_object}{DRUGNEM R object output using down regulation desired priors.}
\item{Noeffect_object}{DRUGNEM R object output using no effect desired priors.}
\item{Upmarker_object}{DRUGNEM R object output using only upregulation desired priors.}
\item{Downmarker_object}{DRUGNEM R object output using only down regulation desired priors.}
\item{Noeffectmarker_object}{DRUGNEM R object output using only no effect desired priors.}
\item{Besteffect_object}{DRUGNEM R object output using the most probable desired effects.}
\item{Rankingindependencemodel}{Table showing the rankings and scores for all drug combinations using no prior desired effects based on independent drug effects.}
\item{Rankingupindependencemodel}{Table showing the rankings and scores for all drug combination using down regulation desired effects data for independence model.}
\item{Rankingdownindependencemodel}{Table showing the rankings and scores for all drug combination using up regulateion desired effects data for independence model.}
\item{Rankingnoeffectindependencemodel}{Table showing the rankings and scores for all drug combination using no effect desired effects data for independence model.}
\item{Rankingupmarkerindependencemodel}{Table showing the rankings and scores for all drug combination using only down desired effects data for independence model.}
\item{Rankingdownmarkerindependencemodel}{Table showing the rankings and scores for all drug combination using only up desired effects data for independence model.}
\item{Rankingnoeffectmarkerindependencemodel}{Table showing the rankings and scores for all drug combination using only no effect desired effects data for independence model.}
\item{Rankingbestindependencemodel}{Table showing the rankings and scores for all drug combination using best desired effects data for independence model.}
\item{Bestdesired_independencemodel}{list of (i) Optimal desired type of effect and (ii) table showing the rankings and scores for all drug combination using the most likely desired effects data for the independence model.}
\item{Bestdesired_nestedmodel}{list of (i) Optimal desired type of effect and (ii) table showing the rankings and scores for all drug combination using the most likely desired effects data for the nested model.}
}
\description{
This algorithm combines Population identification, Nested effects modeling and Rank analysis to identify and rank multi-target drug combinations from single drug effects measured at the level of single cells. Drugnemmain takes as input either (1) a folder containing all drug response FCS files or (2) A data matrix of pooled drug responses for all proteins(columns) across all cells(rows) or (3) a folder with manually gated cells for drug responses from mass cytometry comprising of n drugs and m target (marker) responses on t cells taken from single or  several patient samples and returns several pdf files including both the drug target nested network and the sorted list of all possible drug combinations and their scores with the best drug regimen at the top derived from pre-specified desired effects such as down regulation effects or effects associated with a desired response of a desired marker e.g death marker or from a model with no effects. The algorithm also varies the desired effects, computes the DRUG-NEM score and chooses the model with the desired effects that produce the maximal score.
}
\examples{
drugnemmain(patient,file=input,outputfile=getwd(), manualgates=FALSE,celltype=NULL,drugs=drugs1,drugids=c(1,2,5,7),antibody=antibody1,pProteins=pProteins1,transformlogic=FALSE,asinhp=1,colid=clustids,coln=clustmarkers,rown=NULL,k=clustersize,param=NULL,respondermarker="cPARP",infer="search",fcsdes=NULL,ylabel=ylabel,subanalysis=TRUE,subsamplesize=10000,type="CONTmLL",filter=FALSE,weighttype="downFC",targetprior=TRUE,Fidelity=FALSE,optimizenoeffect=FALSE)
}
