% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsimportmanual.R
\name{fcsimportmanual}
\alias{fcsimportmanual}
\title{A function for importing and density downsampling of several manually gated FCS files from an input directory.}
\usage{
fcsimportmanual(file, drugs, celltype, clustids = NULL,
  transformlogic = FALSE, coln = NULL, rown = NULL, patient,
  subanalysis = FALSE, subsamplesize = NULL)
}
\arguments{
\item{file}{A file path or directory path name with all fcs files of interest.}

\item{drugs}{Parameter vector corresponding to the names of all single drugs. Paremeter must maintain the order and names exactly as in the input folder with the FCS files with the Basal or control file name coming first.}

\item{celltype}{Parameter vector corresponding to the manually gated cell types. Celltype names and ordering must be exact as in the manuall gated file names. Default is NULL if there are no manually gated data.}

\item{clustids}{index of lineage markers of interest to be used by ccast algorithm.}

\item{transformlogic}{An character string that defines the type of transformation. See function read.FCS in R packgae flowCore for more information. Valid values are linearize (default) or scale.The linearize transformation applies the appropriate power transform to the data while the scale transformation scales all columns to $[0,10^decades]$. defaulting to decades=0 as in the FCS4 specification. A logical can also be used: TRUE is equal to linearize and FALSE(which is the default) corresponds to no transformation.}

\item{coln}{column names of markers to use of equal length as clustids. If not provided the default fcs column names will be used.}

\item{rown}{row names of data to use of equal length as number of cells. If not provided the default fcs file names will be used.}

\item{patient}{Patient or subject name. Name must be included in the file names.}

\item{subanalysis}{A logical to downsample original FCS file(s). Default is FALSE. Recommended if total number of cells across all FCS files > 100000.}

\item{subsamplesize}{Number of cells to subsample using density downsampling approach by Qiu et al. 2011}
}
\value{
A FCS file with all cells pooled from all input FCS files and the marker expressions and list of 4 items:
\item{subsetdata }{Dataframe of pooled cell expressions for all selected lineage markers}
\item{allexprsdata }{Dataframe of marker expression of  all pooled cells from original FCS files}
\item{fcsfile}{Flowframe object of all marker expression for all combined cells from original FCS files}
\item{groupids}{List of row ids(cell ids) corresponding to subpopulation of cells}
}
\description{
A function for importing and density downsampling of several manually gated FCS files from an input directory.
}
\examples{
out=fcsimportmanual(file,drugs,celltype,clustids=colid,transformlogic=FALSE,coln=NULL,rown=NULL,patient,subanalysis=FALSE,subsamplesize=subsamplesize)
}
\seealso{
\code{\link{drugmnemmain}} which this function wraps
}
