% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity.R
\name{fidelity}
\alias{fidelity}
\title{A function to identify lineage subpopulations that are affected by drugs.}
\usage{
fidelity(x, drugs, method = "simes", alpha1 = 0.05)
}
\arguments{
\item{x}{List of expression values as data matrices of length equal to the number of lineage subpopulations stored as R object. Each matrix has rows as lineage markers and columns as cells labeled as treatment conditions including Basal.}

\item{method}{character string of either "Simes" (simes) or bonferoni (bonfer).Default is Simes.}

\item{alpha1}{cuttoff p-value parameter for benchmarking adjusted p-values . Default is 0.05.}
}
\value{
A numeric vector of subpopulation ids associated with statistically significant differences across all drugs:
}
\description{
This function takes as input a list of expression values as data matrices of length equal to the number of lineage subpopulations, then applies a simes global test of no change to each subpopulations using all lineage markers and returns the numeric ids for the subpopulations that show statistically significant differences.
}
\examples{
poorid= fidelity(Sdata22,method="simes")

}
\seealso{
\code{\link{comparepopintrasignalalivefidelity}} which uses this function.
}
