% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnemupdownnoeffecttstat.R
\name{fitnemupdownnoeffecttstat}
\alias{fitnemupdownnoeffecttstat}
\title{A function to predict the DRUGNEM network, scores drug combinations and plot diagnostic heatmaps for combination therapy using weighttype=="T-stat" option.}
\usage{
fitnemupdownnoeffecttstat(Out3, Outp, drugs, p = NULL, patient, infer,
  type = "CONTmLL", targetprior)
}
\arguments{
\item{Out3}{R object corresponding to a list of dataframes for each subpopulation with protein expressions for each cell across functional proteins}

\item{Outp}{R object corresponding to a list of dataframes for each subpopulation with Fold change expressions for proteins(rows) and drugs (columns) relative to the basal treatment.}

\item{drugs}{Integer vector of subset index of drug ids to optimize drug combinations from all available drugs.}

\item{p}{Apoptotic marker from the respondermarker parameter. Default is NULL.}

\item{patient}{Patient or subject name. Name must be included exactly as in the file names.}

\item{infer}{Network search approach to score the best drug network. Default is "search" for exhaustive search. For large networks use "nem.greedy"}

\item{type}{Parameter defining the type of data and likelihood for scoring the Nested effect models: Default "CONTmLL" called internally by the nem function in nem package correspond to the effect probabilities while "CONTmLLBayes" correspond to logodds that each protein is differentially expressed.}

\item{targetprior}{A logical to filter out outlier or noisy drug targets. Default is FALSE.}
}
\value{
The output of this function is a list 31 items:
\item{DRUGNEMeffectmodel}{Graphical model showing the nested relationships between drugs and their target proteins using an effect as a desired effect}
\item{Rankingeffectmodel}{Table showing the rankings and scores for all drug combinations using using an effect as a desired effect.}
\item{DRUGNEMupmodel}{Graphical model showing the nested relationships between drugs and their target proteins using effects associated with positive t-stat as desired effects.}
\item{Rankingupmodel}{Table showing the rankings and scores for all drug combination using   effects associated with positive t-stat as desired effects.}
\item{DRUGNEMdownmodel}{Graphical model showing the nested relationships between drugs and their target proteins using  effects associated with negative t-stat as desired effects.}
\item{Rankingdownmodel}{Table showing the rankings and scores for all drug combination using  effects associated with negative t-stat as as desired effects.}
\item{DRUGNEMnoeffectmodel}{Graphical model showing the nested relationships between drugs and their target proteins using  markers associated with no effects as desired effects.}
\item{Rankingnoeffectmodel}{Table showing the rankings and scores for all drug combination using  markers associated with no effects as desired effects.}
\item{DRUGNEMupmarkermodel}{Graphical model showing the nested relationships between drugs and the markers that will produce an upregulation desired effect.}
\item{Rankingupmarkermodel}{Table showing the rankings and scores for all drug combination using the markers that will produce an upregulation desired effect.}
\item{DRUGNEMdownmarkermodel}{Graphical model showing the nested relationships between drugs and the markers that will produce a downregulation desired effect.}
\item{Rankingdownmarkermodel}{Table showing the rankings and scores for all drug combination using the markers that will produce a downregulation desired effect.}
\item{DRUGNEMnoeffectmarkermodel}{Graphical model showing the nested relationships between drugs and the markers that will produce a no effect as desired effects.}
\item{Rankingnoeffectmarkermodel}{Table showing the rankings and scores for all drug combination using the markers that will produce a no effect as desired effects.}
\item{DRUGNEMbesteffectmodel}{Graphical model showing the nested relationships between drugs and their target proteins using most likely observed regulation as desired effects.}
\item{Rankingbesteffectmodel}{Table showing the rankings and scores for all drug combination using most likely observed regulation as desired effects.}
\item{Up_object}{DRUGNEM R object output using positive t-stat as desired priors.}
\item{Down_object}{DRUGNEM R object output using negative t-stat desired priors.}
\item{Noeffect_object}{DRUGNEM R object output using no effect desired priors.}
\item{Upmarker_object}{DRUGNEM R object output using only the markers that will produce an upregulation desired effect.}
\item{Downmarker_object}{DRUGNEM R object output using only the markers that will produce a downregulation desired effect.}
\item{Noeffectmarker_object}{DRUGNEM R object output using only the markers that will produce a no effect as desired effects.}
\item{Besteffect_object}{DRUGNEM R object output using the most likely observed regulation as desired effects.}
\item{Rankingindependencemodel}{Table showing the rankings and scores for all drug combinations using effects associated with positive t-stat as desired effects based on independent drug effects.}
\item{Rankingupindependencemodel}{Table showing the rankings and scores for all drug combination using  effects associated with positive t-stat as desired effects for independence model.}
\item{Rankingdownindependencemodel}{Table showing the rankings and scores for all drug combination using  effects associated with negative t-stat as desired effects for independence model.}
\item{Rankingnoeffectindependencemodel}{Table showing the rankings and scores for all drug combination using no effects as effect desired effects data for independence model.}
\item{Rankingupmarkerindependencemodel}{Table showing the rankings and scores for all drug combination using only the markers that will produce an upregulation desired effect for independence model.}
\item{Rankingdownmarkerindependencemodel}{Table showing the rankings and scores for all drug combination using only the markers that will produce an downregulation desired effect for independence model.}
\item{Rankingnoeffectmarkerindependencemodel}{Table showing the rankings and scores for all drug combination using the markers that will produce an no effect as desired effect for independence model.}
\item{Rankingbestindependencemodel}{Table showing the rankings and scores for all drug combination using the most likely observed regulation as desired effects for independence model.}
}
\description{
A function to predict the DRUGNEM network, scores drug combinations and plot diagnostic heatmaps for combination therapy using weighttype=="T-stat" option.
}
\examples{
res3=fitmnemupdownnoeffecttstat(Out3=Sdata2,Outp=Out2,drugs=dids,p=NULL,patient=patient,infer=infer,type=type)
}
\seealso{
\code{\link{drugnemmain}} which this function wraps
}
