% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnessy2.R
\name{plot.nem2}
\alias{plot.nem2}
\title{plot graph of nested effects model, the marginal likelihood distribution or the posterior position of the effected proteins}
\usage{
plot.nem2(x, what = "graph", remove.singletons = FALSE, PDF = FALSE,
  filename = "nemplot.pdf", thresh = 0, transitiveReduction = FALSE,
  plot.probs = FALSE, SCC = TRUE, D = NULL, draw.lines = FALSE,
  palette = "BlueRed", ...)
}
\arguments{
\item{x}{nem object to plot}

\item{what}{(i), "graph", (ii) "mLL" = likelihood distribution, (iii) "pos" = posterior position of effected proteins}

\item{remove.singletons}{remove unconnected nodes from the graph plot}

\item{PDF}{output as PDF-file}

\item{filename}{filename of PDF-file}

\item{thresh}{if x has a real valued adjacency matrix (weight matrix), don't plot edges with |weight| <= thresh}

\item{transitiveReduction}{plot a transitively reduced graph}

\item{plot.probs}{plot edge weights/probabilities.}

\item{SCC}{plot the strongly connected components graph}

\item{D}{Visualize the nested subset structure of the dataset via plotEffects along with the graph and show the linking of E-proteins to S-proteins in the dataset. Should only be used for small networks. Default: Just plot the graph}

\item{draw.lines}{If the nested subset structure is shown, should additionally lines connecting S-genes and their associated E-genes be drawn? WARNING: For larger datasets than e.g. 5 S-genes this most probably does not work, because the nested subset structure picture then partially overlaps with the graph picture. Default: Do not draw these lines}

\item{palette}{olor palette to use: either 'BlueRed' (default) or 'Grey'}

\item{...}{other arguments to be passed to the Rgraphviz plot function or to the graphics 'image' function.}
}
\value{
None
}
\description{
plot graph of nested effects model, the marginal likelihood distribution or the posterior position of the effected proteins
}
\examples{
plot.nem2(rr,filename=paste(patient,"DRUGMNEM network and Heatmap for all states",infer,type,".pdf",sep=""),main="DRUGMNEM Network",PDF=TRUE,what = "graph",D=cdata1,draw.lines = FALSE)
}
\seealso{
\code{\link{fitmnemdown}} which this function wraps
}
