% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responders.R
\name{responders}
\alias{responders}
\title{A function to filter out death or dying cells and return responder cell ids.}
\usage{
responders(rowids, dataP, drugs = drugs, asinhp = asinhp, colid = colid,
  antibody = antibody, patient = patient, outfile = dirc,
  respondermarker = NULL)
}
\arguments{
\item{rowids}{List of row ids(cell ids) for each sub population of cells stored as R object.}

\item{dataP}{Marker expression data matrix of pooled cells from original FCS file stored as R object.}

\item{drugs}{Parameter vector corresponding to the names of all single drugs.}

\item{asinhp}{asinh parameter p for scaling the data if necessary. Default is 1.}

\item{colid}{vector apoptotic marker ids of interest to be used to define responders and nonresponders}

\item{antibody}{Vector of marker names of length equal to length of the columns in each FCS file.}

\item{patient}{Patient or subject name. Name must be included exactly as in the file names.}

\item{outfile}{Name of working directory. Drugmenem will use this directory to create other subdirectories for filtering analysis results.}

\item{respondermarker}{Vector string of apoptotic markers to specified by user. Must contain atleast one element. Default is NULL}
}
\value{
List of dataframes with mean and fold change expression values for each drug relative to the baseline and heatmaps of fold changes with relative percentage of cells compared to basal for all drugs. The output of this functionis a list 3 items:
\item{SensitiveCCASTanalysis}{ccast output object containing a list of 8 items with the last corresponding to the responder and non-responder cells derived from the input FCS fileid}
\item{pooledresponders}{List of cell ids of length equal to the number of cell types corresponding to responders}
\item{poolednonresponders}{List of cell ids of length equal to the number of cell types corresponding to responders.}
}
\description{
A function to filter out death or dying cells and return responder cell ids.
}
\examples{
res11=responders(rowids,dataP,drugs=drugs,asinhp=asinhp,colid=colid,antibody=antibody,patient=patient,outfile=dirc,respondermarker=respondermarker)
}
\seealso{
\code{\link{drugmnemmain}} which this function wraps
}
