% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE.addDensity.downsample.R
\name{SPADE.addDensity.downsample}
\alias{SPADE.addDensity.downsample}
\title{Function to perform density downsampling of a given FCS file}
\usage{
SPADE.addDensity.downsample(infilename, cols = colid,
  arcsinh_cofactor = asinhp, kernel_mult = 5, apprx_mult = 1.5,
  med_samples = 2000, comp = TRUE, exclude_pctile = 0.01,
  target_pctile = 0.05, desired_samples = subsamplesize)
}
\arguments{
\item{infilename}{String of FCS file name that should be used as input}

\item{cols}{vector of column ids of lineage markers of interest to be used by ccast algorithm to define cell clusters.}

\item{arcsinh_cofactor}{Cofactor used in arcsinh transform asinh(data/arcsinh_cofactor) of data}

\item{kernel_mult}{Multiplier of the minimum median distance within which other observations are counted towards the density}

\item{apprx_mult}{Multiplier of the minimum median distance within which observations are approximated to have the same density}

\item{med_samples}{Number of observations used to estimate the minimum median distance}

\item{comp}{Apply compensation matrix if present in SPILL or SPILLOVER keywords}

\item{exclude_pctile}{Numeric value in [0,1]. Densities below this percentile will be excluded.}

\item{target_pctile}{Numeric value in [0,1]. Densities below this percentile, but above exclude_pctile will be retained. Only meaningful if desired_samples is NULL.}

\item{desired_samples}{Desired number of samples. If set to integer value, the target percentile will be set internally to downsample to approximately the desired number of samples.}
}
\value{
The name of the written file is returned in addition to a dataframe of subsampled cells with their marker expression values.
}
\description{
Function to perform density downsampling of a given FCS file
}
\examples{

}
\seealso{
\code{\link{fcsimport3}} which this function wraps
}
